Moved worker into tests/utils
